+++ Classes/Domain/Service/QualificationService.php
@@ -91,6 +91,7 @@ class QualificationService
                 // Check if previous qualification exist
                 $histories = new \TYPO3\CMS\Extbase\Persistence\ObjectStorage();
                 $prevStatus = FALSE;
+                /* @var $prevEmployeequalification \Pmwebdesign\Staffm\Domain\Model\Employeequalification */
                 foreach ($prevEmployeequalifications as $prevEmployeequalification) {
                     if ($prevEmployeequalification->getQualification() === $qualification) {
                         $prevStatus = TRUE;
@@ -113,6 +114,7 @@ class QualificationService
                                     // Set new status
                                     $prevEmployeequalification->setStatus($status);
                                 }
+                                
                                 // Add new history
                                 $newHistory = new \Pmwebdesign\Staffm\Domain\Model\History();
                                 $newHistory->setStatus($status);
@@ -122,6 +124,24 @@ class QualificationService
                                 $histories->attach($newHistory);
                                 
                                 $prevEmployeequalification->setHistories($histories);
+                            // Same status and current user is the same as history user?
+                            } else if ($prevEmployeequalification->getStatus() == $status && count($histories) > 0) {
+                                // Get last history
+                                $sum = $histories->count();
+                                $counter = 1;          
+                                foreach ($histories as $history) {
+                                    if($sum == $counter) {
+                                        /* @var $lastHistory \Pmwebdesign\Staffm\Domain\Model\History */      
+                                        $lastHistory = $history;
+                                    }
+                                    $counter++;
+                                }       
+                                
+                                // Current user is the same as history user?
+                                if($lastHistory->getAssessor() === $assessor) {
+                                    // Update notice
+                                    $lastHistory->setNote($note);                                   
+                                }
                             }
                         }
                         if ($note != null) {